libcopybit: fix for narrowing conversion in C++11.

+++ libcopybit/copybit_c2d.cpp
@@ -1361,8 +1361,8 @@ static int blit_copybit(
     struct copybit_image_t const *src,
     struct copybit_region_t const *region)
 {
-    struct copybit_rect_t dr = { 0, 0, dst->w, dst->h };
-    struct copybit_rect_t sr = { 0, 0, src->w, src->h };
+    struct copybit_rect_t dr = { 0, 0, static_cast<int>(dst->w), static_cast<int>(dst->h) };
+    struct copybit_rect_t sr = { 0, 0, static_cast<int>(src->w), static_cast<int>(src->h) };
     return stretch_copybit_internal(dev, dst, src, &dr, &sr, region, false);
 }
